requirements.txt
Flask==2.3.3
Flask_SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
